Migrate middle-out code into Graphite vert module.
This will allow the SkSL compiler to avoid reprocessing it each time
it is used. This reduces the input SkSL for most Graphite vertex
programs to just a few lines of code.

This also deduplicates a lot of identical SkSL between the
TesselateWedges and TesselateCurves render steps.

Change-Id: I697b3284eba1ede8a2643ba9f4d76b699d3fd756
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/533576
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>