[bazel] Use minimal download flag when building on CI
https://bazel.build/reference/command-line-reference#flag--remote_download_minimal

This will hopefully speed up CI builds by a few seconds
because we don't have to copy the builds from the
RBE instances to our machines running the bazel task.

Testing locally, a build with an empty local bazel cache
but warm remote cache took 35 seconds with the status
quo and 13 seconds with the new flag.

Change-Id: I04ba4b0d7ba7864746c52d103be14f4f7b238d39
Bug: skia:12541
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/533756
Reviewed-by: Erik Rose <erikrose@google.com>
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>