[xml] Correct assert that parser is in an element.
In SkDOM the predicate "is currently inside an element" is expressed by
`fLevel > 0`. When starting an element this should be tested before
incrementing fLevel since this is testing if the element being started
is top level or not. When ending an element this should be tested before
decrementing fLevel since when ending an element the parser is currently
inside an element.

Previously the decrement when ending an element occured too soon,
causing flushAttributes to fail a pre-condition assert that it is not
run at top level (without an element present). Fix this by moving the
decrement (which may signal a return to the top level) after the call to
flushAttributes in onEndElement.

Change-Id: I1a45a3dae5a5af97ca3340e56ddc18b9ae94d374
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/533808
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>