Isolate the buffers for generating SubRuns
Isolate the buffers to move them from the Painter, where they are
situated on the Device(CPU) or SurfaceDrawContext(GPU), to the
Canvas. This will allow better sharing of the buffers, and the
splitting out of the CPU and GPU parts of the Painter.

Change-Id: Ie2a99dc7f9e784c4f00012b71ece5ff482687a88
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/538660
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>