Optional OT-SVG support for DirectWrite.
This does not yet support gzip'ed svgDoc. DirectWrite just hands back
the raw svgDoc data here which may be in any format. It does not attempt
to check for the gzip header or flate the contents. Skia will eventually
need to do this itself for full support. At the moment glyphs which are
backed by a gzip'ed svgDoc will fall back to outlines.

Change-Id: I34a92b2285189ecdbdbc6a8b2a668bd4935bca15
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/541061
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>