[pdf] Reduce nesting in gradient functions.
Previously the function emitted for complex gradients looked like

    if (t <= 0) {
      ret = color[0]
    } else {
      if (t <= stop[1]) {
        ret = interp(t, color[0], color[1])
      } else {
        if (t <= stop[2]) {
          ret = interp(t, color[1], color[2])
        } else {
          if (t <= stop[3]) {
            ret = interp(t, color[2], color[3])
          } else {
            if (t <= stop[4]) {
              ret = interp(t, color[3], color[4])
            } else {
              ret = color[4]
            }
          }
        }
      }
    }

So a gradient with many color stops could end up with extreme nesting of
ifelse. This could cause some viewers to reject the function due to
implementation limits. Even when not rejected the drawing would be slow
due to the linear testing of each range.

Change the function's structure to a binary search with special cases
for the begin and end ranges.

    if (t <= 0) {
      ret = color[0];
      t = 0;
    }
    if (t > 0 && t <= stop[4]) {
      if (t <= stop[2]) {
        if (t <= stop[1]) {
          ret = interp(t, color[0], color[1]);
        } else {
          ret = interp(t, color[1], color[2]);
        }
      } else {
        if (t <= stop[3] {
          ret = interp(t, color[2], color[3]);
        } else {
          ret = interp(t, color[3], color[4]);
        }
      }
      t = 0;
    }
    if (t > 0) {
      ret = color[4];
    }

This change allows gradients with many stops to work in all known
viewers and greatly speeds up their drawing in viewers which did manage
to render with the old function. The function text is slightly larger
with this change, but minimally. If this becomes a size issue it may be
possible to size optimize gradients with few stops in the future.

Props to Bram De Cooman for suggesting this form for the function, the
initial implementation [0], and testing to make sure this approach would
resolve the issue.

[0] https://skia-review.googlesource.com/c/skia/+/540136

Bug: chromium:1316377
Co-authored-by: Bram De Cooman <bram.decooman@gmail.com>
Change-Id: Ib29cbb17220b9c32105b02222c1e1615ce0196bf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/541639
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Herb Derby <herb@google.com>